@@ -1,6 +1,6 @@
 export class Todo {
-  id: string;
-  title: string;
-  completed: boolean;
-  createdAt: Date;
+  id: string = '';
+  title: string = '';
+  completed: boolean = false;
+  createdAt: Date = new Date();
 }
